requirements.txt
implicit
Flask
numpy
pandas
scikit-learn
mysql_connector-python
sqlalchemy